[feat] : @ConditionOnExpression 추가

+++ Ticket-Api/src/main/java/com/jnu/ticketapi/api/registration/service/RegistrationUseCase.java
@@ -30,13 +30,16 @@ import com.jnu.ticketdomain.domains.user.adaptor.UserAdaptor;
 import com.jnu.ticketdomain.domains.user.domain.User;
 import com.jnu.ticketdomain.domains.user.domain.UserStatus;
 import com.jnu.ticketinfrastructure.redis.RedisService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.transaction.annotation.Transactional;
+
 import java.time.LocalDateTime;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.transaction.annotation.Transactional;
 
 @UseCase
 @RequiredArgsConstructor
@@ -50,8 +53,11 @@ public class RegistrationUseCase {
     private final EventWithDrawUseCase eventWithDrawUseCase;
     private final Encryption encryption;
     private final ValidateCaptchaUseCase validateCaptchaUseCase;
-    private final RedisService redisService;
+    @Autowired(required = false) private RedisService redisService;
     private final EventAdaptor eventAdaptor;
+
+    @Value("${ableRedis:true}")
+    private boolean ableRedis;
 
     public Registration saveAndFlush(Registration registration) {
         return registrationAdaptor.saveAndFlush(registration);
@@ -163,7 +169,9 @@ public class RegistrationUseCase {
             Long eventId) {
         tempRegistration.update(registration);
         eventWithDrawUseCase.issueEvent(registration, user.getId(), sectorId, eventId);
-        redisService.deleteValues("RT(" + TicketStatic.SERVER + "):" + email);
+        if (ableRedis) {
+            redisService.deleteValues("RT(" + TicketStatic.SERVER + "):" + email);
+        }
     }
 
     private FinalSaveResponse saveRegistration(
@@ -184,7 +192,9 @@ public class RegistrationUseCase {
             Long eventId) {
         //        Registration saveReg = saveAndFlush(registration);
         eventWithDrawUseCase.issueEvent(registration, currentUserId, sector.getId(), eventId);
-        redisService.deleteValues("RT(" + TicketStatic.SERVER + "):" + email);
+        if (ableRedis) {
+            redisService.deleteValues("RT(" + TicketStatic.SERVER + "):" + email);
+        }
         //        Events.raise(new EventIssuedEvent())
         return FinalSaveResponse.from(registration);
     }
@@ -229,8 +239,8 @@ public class RegistrationUseCase {
                                         .thenComparing(
                                                 r ->
                                                         r.getUser()
-                                                                        .getStatus()
-                                                                        .equals(UserStatus.SUCCESS)
+                                                                .getStatus()
+                                                                .equals(UserStatus.SUCCESS)
                                                                 ? r.getId()
                                                                 : r.getUser().getSequence())
                                         .thenComparing(
